# server/docs/swagger.yaml
definitions:
  main.user:
    properties:
      department:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      userStatus:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /users:
    get:
      description: Retrieve a list of users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.user'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a user with the provided information
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.user'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.user'
        "400":
          description: Bad Request
          schema: {}
        "409":
          description: Conflict
          schema: {}
      summary: Add a new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Remove a user from the database by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: User deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
      summary: Delete a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user information by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.user'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.user'
        "400":
          description: Bad Request
          schema: {}
      summary: Update an existing user
      tags:
      - users
swagger: "2.0"